requirements.txt
amqp==5.1.1
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.1
Automat==22.10.0
billiard==4.1.0
binance-connector==3.5.1
celery==5.3.0
certifi==2023.11.17
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
constantly==15.1.0
cron-descriptor==1.4.0
cryptography==41.0.1
daphne==4.0.0
decorator==5.1.1
Django==4.2.2
django-celery-beat==2.5.0
django-cors-headers==4.0.0
django-filter==23.2
django-ipware==5.0.0
django-model-utils==4.3.1
django-timezone-field==5.0
djangorestframework==3.14.0
dnspython==2.4.2
executing==2.0.1
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipython==8.21.0
jedi==0.19.1
kombu==5.3.0
matplotlib-inline==0.1.6
msgpack==1.0.5
numpy==1.26.2
pandas==2.1.4
parso==0.8.3
Pillow==9.5.0
prompt-toolkit
psycopg==3.1.9
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.0
Pygments==2.17.2
pymongo==4.6.0
pyOpenSSL==23.2.0
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2023.3
redis==4.5.5
requests==2.31.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.14.1
Twisted==22.10.0
# twisted-iocpsupport==1.0.3
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.1.0
vine==5.0.0
wcwidth==0.2.6
websocket-client==1.7.0
whitenoise==6.4.0
zope.interface==6.0
